+++ rollup.config.js
@@ -9,45 +9,45 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const json = require("@rollup/plugin-json");
-const commonjs = require("@rollup/plugin-commonjs");
-const resolve = require("@rollup/plugin-node-resolve");
-const babel = require("rollup-plugin-babel");
-const uglifier = require("rollup-plugin-uglify");
+import json from "@rollup/plugin-json";
+import commonjs from "@rollup/plugin-commonjs";
+import resolve from "@rollup/plugin-node-resolve";
+import babel from "@rollup/plugin-babel";
+import { terser } from "rollup-plugin-terser";
 
 const BASE_PLUGINS = [
   resolve({
     mainFields: ["module", "main"],
-    preferBuiltins: true
+    preferBuiltins: true,
   }),
   commonjs({
-    include: "node_modules/**"
+    include: "node_modules/**",
   }),
   json(),
   babel({
     babelrc: false,
-    presets: ["@babel/env"]
-  })
+    presets: ["@babel/env"],
+  }),
 ];
 
 const BASE_CONFIG = {
   input: "src/index.js",
   external: ["redux", "is-promise"],
-  plugins: [...BASE_PLUGINS, uglifier.uglify()]
+  plugins: [...BASE_PLUGINS, terser()],
 };
 
 const GLOBALS = {
   redux: "Redux",
-  "is-promise": "isPromise"
+  "is-promise": "isPromise",
 };
 
-module.exports = [
+export default [
   {
     ...BASE_CONFIG,
     output: {
       file: "dist/index.cjs.js",
-      format: "cjs"
-    }
+      format: "cjs",
+    },
   },
   {
     ...BASE_CONFIG,
@@ -55,16 +55,16 @@ module.exports = [
       file: "dist/index.umd.js",
       format: "umd",
       name: "dataProvider",
-      globals: GLOBALS
-    }
+      globals: GLOBALS,
+    },
   },
   {
     ...BASE_CONFIG,
     output: {
       file: "dist/index.esm.js",
       format: "esm",
-      globals: GLOBALS
+      globals: GLOBALS,
     },
-    plugins: BASE_PLUGINS
-  }
+    plugins: BASE_PLUGINS,
+  },
 ];