+++ test-e2e/browser/react-app/src/modules/authors/Authors.js
@@ -6,7 +6,8 @@ import AuthorsList from "modules/authors-list";
 import SectionContainer from "components/section-container";
 import ItemsTitle from "components/items-title";
 import ItemsListContainer from "components/items-list-container";
-import { useRefresh, useData, useLoading } from "helpers/data-provider";
+import Placeholders from "components/placeholders";
+import { useRefresh, useData, useLoading, useLoaded } from "helpers/data-provider";
 
 import AuthorNew from "./modules/author-new";
 
@@ -14,15 +15,20 @@ const Authors = () => {
   useRefresh(authorsProvider);
   const authors = useData(authorsProvider);
   const loading = useLoading(authorsProvider);
+  const loaded = useLoaded(authorsProvider);
 
   console.log("Rendering authors", loading, authors);
 
   return (
     <SectionContainer loading={loading} id="authors-column">
       <ItemsTitle title="Authors" />
-      <ItemsListContainer id="authors-container">
-        <AuthorsList authors={authors} />
-      </ItemsListContainer>
+      {loaded ? (
+        <ItemsListContainer id="authors-container">
+          <AuthorsList authors={authors} />
+        </ItemsListContainer>
+      ) : (
+        <Placeholders id="authors-placeholder" />
+      )}
       <AuthorNew />
     </SectionContainer>
   );