+++ test-e2e/browser/react-app/src/modules/books/Books.js
@@ -6,7 +6,8 @@ import BooksList from "modules/books-list";
 import SectionContainer from "components/section-container";
 import ItemsTitle from "components/items-title";
 import ItemsListContainer from "components/items-list-container";
-import { useRefresh, useData, useLoading } from "helpers/data-provider";
+import Placeholders from "components/placeholders";
+import { useRefresh, useData, useLoading, useLoaded } from "helpers/data-provider";
 
 import BookNew from "./modules/book-new";
 
@@ -14,15 +15,20 @@ const Books = () => {
   useRefresh(booksWithAuthorName);
   const books = useData(booksWithAuthorName);
   const loading = useLoading(booksWithAuthorName);
+  const loaded = useLoaded(booksWithAuthorName);
 
-  console.log("Rendering books", loading, books);
+  console.log("Rendering books", loading, books, loaded);
 
   return (
     <SectionContainer loading={loading} id="books-column">
       <ItemsTitle title="Books" />
-      <ItemsListContainer id="books-container">
-        <BooksList books={books} />
-      </ItemsListContainer>
+      {loaded ? (
+        <ItemsListContainer id="books-container">
+          <BooksList books={books} />
+        </ItemsListContainer>
+      ) : (
+        <Placeholders id="books-placeholder" />
+      )}
       <BookNew />
     </SectionContainer>
   );